push_swap[ci]: refactore ft_sort_stack() add ft_get_sm_value() ft_get_bg_value(). Create new files to organize the project: list_management.c, list_utils.c, sort_functions.c
