+++ push_swap/push_swap.c
@@ -72,98 +72,6 @@ t_stack	*ft_handle_args(t_stack **stack_a, int argc, char **argv)
 	return (*stack_a);
 }
 
-t_stack	*ft_init_lst(int value)
-{
-	t_stack	*lst;
-
-	lst = malloc (sizeof(t_stack));
-	if (!lst)
-		return (NULL);
-	lst -> index = 0;
-	lst -> value = value;
-	lst -> next = lst;
-	lst -> previous = lst;
-	return (lst);
-}
-
-t_stack	*ft_push_lst(t_stack **stack, t_stack *new)
-{
-	t_stack *last;
-
-	if (!new)
-		return (NULL);
-	if (!*stack)
-		*stack = new;
-	else
-	{
-		new -> index = ((*stack) -> previous -> index) + 1;
-		last = (*stack) -> previous;	
-		last -> next = new;
-		new -> previous = last;
-		new -> next = *stack;
-		(*stack) -> previous = new;
-	}
-	return (*stack);
-}
-
-t_stack	*ft_unshift_lst(t_stack **stack, t_stack *new)
-{
-	t_stack	*last;
-
-	if (!new)
-		return (NULL);
-	if (*stack)
-	{
-		last = (*stack) -> previous;
-		last -> next = new;
-		(*stack) -> previous = new;
-		new -> next = *stack;
-		new -> previous = last;
-		ft_update_stack_index(new);
-	}
-	*stack = new;
-	return (*stack);
-}
-
-t_stack	*ft_get_lst_by_index(t_stack *stack, size_t index)
-{
-	t_stack	*lst;
-
-	lst = stack;
-	if (!stack)
-		return (NULL);
-	while (lst -> index != index)
-		lst = lst -> next;
-	return (lst);
-}
-
-void	ft_free_stack(t_stack *stack)
-{
-	t_stack	*tmp;
-	t_stack	*curr;
-
-	printf("FUNCTION FREE STACK\n");
-	curr = stack;
-	if (stack)
-	{
-		while (curr -> next != stack)
-		{
-			printf("///////////////////////===>\n");
-			tmp = curr;
-			printf("FREE stack -> index: |%ld|\n", tmp -> index); 
-			printf("FREE stack -> value: |%d|\n", tmp -> value); 
-			curr = curr -> next;
-			free(tmp);
-			tmp = NULL;
-		}
-		printf("///////////////////////===>\n");
-		printf("+FREE stack -> index: |%ld|\n", curr -> index); 
-		printf("+FREE stack -> value: |%d|\n", curr -> value); 
-		printf("///////////////////////===>\n");
-	}
-	free(curr);	
-	curr = NULL;
-}
 
 
 /*
